# lib/api-zod/src/generated/types/jobSourceInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * VacancyMatch AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JobSourceInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  sourceType: string;
  baseUrl?: string;
  isActive?: boolean;
}